requirements.txt
selenium
webdriver-manager
Pillow
